requirements.txt
appdirs==1.4.3
coreapi==2.3.0
coreschema==0.0.4
dateutils==0.6.6
defusedxml==0.5.0
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.6.3
djangorestframework-xml==1.3.0
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
